requirements.txt
beautifulsoup4==4.11.1
discord.py==2.3.0
humanize==4.6.0
pafy==0.5.5
psutil==5.9.5
python-dotenv==1.0.0
Requests==2.31.0
spotipy==2.23.0
youtube_dl==2021.12.17
youtube_search_python==1.6.6
